requirements.txt
fastmcp==2.12.3
pyyaml
pydantic
python-dotenv
lxml
pathspec
tenacity
markupsafe
